test(tui): budget test + file splits enforcing 800-LOC cap (#1018)
