+++ sdk/tui/src/app.rs
@@ -8,314 +8,21 @@
 // OF ANY KIND, either express or implied. See the License for the specific language
 // governing permissions and limitations under the License.
 
-use std::collections::HashSet;
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 
 use crossterm::event::{KeyCode, KeyEvent, KeyModifiers, MouseButton, MouseEvent, MouseEventKind};
-use design_data_core::cascade::{self, ResolutionContext, specificity};
-use design_data_core::diff::display_name;
-use design_data_core::graph::{Layer, TokenGraph, TokenRecord};
-use design_data_core::query;
-use design_data_core::schema::SchemaRegistry;
-use design_data_core::validate;
+use design_data_core::cascade::ResolutionContext;
 use ratatui::layout::Rect;
 use ratatui::widgets::TableState;
 use tui_input::Input;
 use tui_input::backend::crossterm::EventHandler;
 
-use crate::find::{FindEvent, FindWizardState};
-use crate::naming::{NamingEvent, NamingWizardState};
-use crate::wizard::{WizardCtx, WizardEvent, WizardState};
+use crate::find::FindEvent;
+use crate::naming::NamingEvent;
+use crate::wizard::{WizardCtx, WizardEvent};
 use crate::wizard_draft::{clear_wizard_draft, from_draft, load_wizard_draft};
 
-/// Command names for Tab autocomplete.
-pub(crate) const KNOWN_COMMANDS: &[&str] =
-    &["find", "name", "new", "query", "resolve", "describe", "validate"];
-
-/// Max palette history entries persisted to disk.
-pub(crate) const HISTORY_CAP: usize = 200;
-
-/// Which prefix the palette was opened with.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum PaletteMode {
-    /// `:` — explicit command mode.
-    Command,
-    /// `/` — fuzzy-find mode.
-    FuzzyFind,
-}
-
-/// Severity of a status bar message; controls render colour.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum StatusKind {
-    Info,
-    Error,
-}
-
-/// A status bar message with its display kind.
-#[derive(Debug, Clone)]
-pub struct StatusMessage {
-    pub text: String,
-    pub kind: StatusKind,
-}
-
-impl StatusMessage {
-    pub fn info(text: impl Into<String>) -> Self {
-        Self { text: text.into(), kind: StatusKind::Info }
-    }
-    pub fn error(text: impl Into<String>) -> Self {
-        Self { text: text.into(), kind: StatusKind::Error }
-    }
-}
-
-// ── View state types ──────────────────────────────────────────────────────────
-
-/// One row in the query results table.
-#[derive(Debug, Clone)]
-pub struct QueryRow {
-    pub name: String,
-    pub value: String,
-    pub file: String,
-    pub layer: String,
-}
-
-impl QueryRow {
-    pub(crate) fn from_record(t: &TokenRecord) -> Self {
-        let value = t
-            .raw
-            .get("value")
-            .map(|v| {
-                if v.is_string() {
-                    v.as_str().unwrap_or("").to_string()
-                } else {
-                    v.to_string()
-                }
-            })
-            .or_else(|| t.alias_target.clone())
-            .unwrap_or_default();
-        let file = t.file.file_name().map(|f| f.to_string_lossy().into_owned()).unwrap_or_default();
-        Self { name: display_name(t), value, file, layer: layer_str(t.layer).to_string() }
-    }
-}
-
-/// State for an active query view.
-pub struct QueryView {
-    pub expr_text: String,
-    pub rows: Vec<QueryRow>,
-    pub table_state: TableState,
-}
-
-impl QueryView {
-    pub fn new(expr_text: String, rows: Vec<QueryRow>) -> Self {
-        let mut table_state = TableState::default();
-        if !rows.is_empty() {
-            table_state.select(Some(0));
-        }
-        Self { expr_text, rows, table_state }
-    }
-
-    pub(crate) fn selected_row(&self) -> Option<&QueryRow> {
-        self.table_state.selected().and_then(|i| self.rows.get(i))
-    }
-}
-
-/// One row in the resolve candidates table.
-#[derive(Debug, Clone)]
-pub struct ResolvedRow {
-    pub name: String,
-    pub value: String,
-    pub file: String,
-    pub layer: String,
-    pub specificity: u32,
-    pub is_winner: bool,
-}
-
-/// State for a resolve results view (winner + ranked candidates).
-pub struct ResolveView {
-    pub property: String,
-    pub rows: Vec<ResolvedRow>,
-    pub table_state: TableState,
-}
-
-impl ResolveView {
-    pub(crate) fn new(property: String, rows: Vec<ResolvedRow>) -> Self {
-        let mut table_state = TableState::default();
-        if !rows.is_empty() {
-            table_state.select(Some(0));
-        }
-        Self { property, rows, table_state }
-    }
-
-    pub(crate) fn selected_row(&self) -> Option<&ResolvedRow> {
-        self.table_state.selected().and_then(|i| self.rows.get(i))
-    }
-}
-
-/// State for a component describe view.
-pub struct DescribeView {
-    pub component: String,
-    pub pretty_json: String,
-    pub scroll: u16,
-}
-
-/// One row in the validate diagnostics table.
-#[derive(Debug, Clone)]
-pub struct DiagnosticRow {
-    pub severity: String,
-    pub rule_id: String,
-    pub token: String,
-    pub message: String,
-}
-
-/// State for a validate findings view.
-pub struct ValidateView {
-    pub rows: Vec<DiagnosticRow>,
-    pub table_state: TableState,
-}
-
-impl ValidateView {
-    pub(crate) fn new(rows: Vec<DiagnosticRow>) -> Self {
-        let mut table_state = TableState::default();
-        if !rows.is_empty() {
-            table_state.select(Some(0));
-        }
-        Self { rows, table_state }
-    }
-
-    pub(crate) fn selected_row(&self) -> Option<&DiagnosticRow> {
-        self.table_state.selected().and_then(|i| self.rows.get(i))
-    }
-}
-
-/// Which view the active area is showing.
-pub enum ActiveView {
-    Empty,
-    Query(QueryView),
-    Resolve(ResolveView),
-    Describe(DescribeView),
-    Validate(ValidateView),
-}
-
-// ── Modals ────────────────────────────────────────────────────────────────────
-
-/// State for the `?` help overlay.
-pub struct HelpModal {
-    pub scroll: u16,
-}
-
-/// An overlay modal that temporarily captures all keyboard input.
-pub enum Modal {
-    Find(Box<FindWizardState>),
-    Wizard(Box<WizardState>),
-    Naming(Box<NamingWizardState>),
-    Help(HelpModal),
-}
-
-impl Modal {
-    /// Whether mouse-wheel scroll events should be routed into this modal.
-    ///
-    /// Only `Wizard` (diff preview) and `Help` have scrollable content.
-    /// New modals default to `false`; override by adding a variant here.
-    pub fn wants_scroll(&self) -> bool {
-        matches!(self, Modal::Wizard(_) | Modal::Help(_))
-    }
-
-    /// Route a scroll delta into this modal's scrollable region.
-    ///
-    /// Only called when `wants_scroll()` returns `true`.
-    pub fn on_scroll(&mut self, delta: i32) {
-        match self {
-            Modal::Wizard(ws) => apply_scroll_delta(&mut ws.diff_scroll, delta),
-            Modal::Help(hm) => apply_scroll_delta(&mut hm.scroll, delta),
-            Modal::Find(_) | Modal::Naming(_) => {}
-        }
-    }
-
-    /// Persist any in-progress state to disk (no-op for modals without persistence).
-    pub fn persist(&self) {
-        use crate::wizard_draft::{save_wizard_draft, to_draft};
-        if let Modal::Wizard(ws) = self {
-            save_wizard_draft(&to_draft(ws));
-        }
-    }
-
-    /// One-line breadcrumb for the current screen, e.g. `"Step 1 of 2 — Filters"`.
-    ///
-    /// Intended for a future status-line indicator that shows which modal is open and
-    /// which screen the user is on.  Not yet wired to a renderer.
-    pub fn screen_label(&self) -> String {
-        match self {
-            Modal::Find(fs) => {
-                use crate::find::FindScreen;
-                let (n, name) = match fs.screen {
-                    FindScreen::Filters => (1u8, "Filters"),
-                    FindScreen::Preview => (2u8, "Preview"),
-                };
-                format!("Step {n} of 2 — {name}")
-            }
-            Modal::Naming(ns) => {
-                use crate::naming::NamingScreen;
-                let (n, name) = match ns.screen {
-                    NamingScreen::Intent => (1u8, "Intent"),
-                    NamingScreen::Classification => (2u8, "Classification"),
-                    NamingScreen::Result => (3u8, "Result"),
-                };
-                format!("Step {n} of 3 — {name}")
-            }
-            Modal::Wizard(ws) => {
-                use crate::wizard::WizardScreen;
-                let (n, total, name) = match ws.screen {
-                    WizardScreen::Intent => (1u8, 4u8, "Intent"),
-                    WizardScreen::Classification => (2, 4, "Classification"),
-                    WizardScreen::Values => (3, 4, "Values"),
-                    WizardScreen::Confirm => (4, 4, "Confirm"),
-                };
-                format!("Step {n} of {total} — {name}")
-            }
-            Modal::Help(_) => "Help".to_string(),
-        }
-    }
-}
-
-// ── Hit regions (mouse support) ───────────────────────────────────────────────
-
-/// What clicking a region does.
-pub enum HitAction {
-    /// Selects a row in the active list or table view.
-    SelectListRow(usize),
-}
-
-/// A rectangular region on screen with an associated action and text content.
-pub struct HitRegion {
-    pub rect: Rect,
-    pub action: HitAction,
-    /// Text representation of this element, used for drag-select copy.
-    pub text: String,
-}
-
-// ── Submit context ────────────────────────────────────────────────────────────
-
-/// Context passed to `submit_palette`; carries the graph plus optional paths for
-/// describe and validate commands.
-pub struct SubmitContext<'a> {
-    pub graph: &'a TokenGraph,
-    pub dataset_path: Option<&'a Path>,
-    pub components_dir: Option<&'a Path>,
-    pub schema_registry: Option<&'a SchemaRegistry>,
-    pub mode_sets_dir: Option<&'a Path>,
-}
-
-impl<'a> SubmitContext<'a> {
-    /// Minimal context for tests and use-cases that only need `:query`.
-    pub fn new(graph: &'a TokenGraph) -> Self {
-        Self {
-            graph,
-            dataset_path: None,
-            components_dir: None,
-            schema_registry: None,
-            mode_sets_dir: None,
-        }
-    }
-}
+pub use crate::app_views::*;
 
 // ── App ───────────────────────────────────────────────────────────────────────
 
@@ -838,306 +545,8 @@ impl App {
         }
     }
 
-    // ── Palette dispatch ─────────────────────────────────────────────────────
-
-    /// Dispatch a committed palette command against the graph and optional context paths.
-    pub fn submit_palette(&mut self, ctx: &SubmitContext<'_>) {
-        if self.palette_mode != PaletteMode::Command {
-            self.palette_open = false;
-            self.palette_input = Input::default();
-            return;
-        }
-
-        let raw = self.palette_input.value().trim().to_string();
-        self.palette_open = false;
-        self.palette_input = Input::default();
-        self.palette_history_cursor = None;
-
-        // Append to history (dedupe head, cap at HISTORY_CAP).
-        if !raw.is_empty()
-            && self.palette_history.first().map(|s| s.as_str()) != Some(raw.as_str())
-        {
-            self.palette_history.insert(0, raw.clone());
-            self.palette_history.truncate(HISTORY_CAP);
-            save_palette_history(&self.palette_history);
-        }
-
-        let (cmd, rest) = match raw.split_once(' ') {
-            Some((c, r)) => (c.to_lowercase(), r.trim().to_string()),
-            None => (raw.to_lowercase(), String::new()),
-        };
-
-        match cmd.as_str() {
-            "query" => {
-                if rest.is_empty() {
-                    self.status_message = Some(StatusMessage::error("query: expression required"));
-                    return;
-                }
-                match query::parse(&rest) {
-                    Ok(expr) => {
-                        let records = query::filter(ctx.graph, &expr);
-                        let rows: Vec<QueryRow> =
-                            records.iter().map(|r| QueryRow::from_record(r)).collect();
-                        let count = rows.len();
-                        self.active_view = ActiveView::Query(QueryView::new(rest.clone(), rows));
-                        self.status_message =
-                            Some(StatusMessage::info(format!("{count} token(s) matched")));
-                    }
-                    Err(e) => {
-                        self.status_message =
-                            Some(StatusMessage::error(format!("query error: {e}")));
-                    }
-                }
-            }
-            "resolve" => {
-                if rest.is_empty() {
-                    self.status_message =
-                        Some(StatusMessage::error("resolve: property=<name> required"));
-                    return;
-                }
-                let (prop, res_ctx) = match parse_resolve_args(&rest) {
-                    Ok(v) => v,
-                    Err(e) => {
-                        self.status_message =
-                            Some(StatusMessage::error(format!("resolve: {e}")));
-                        return;
-                    }
-                };
-                let candidates: Vec<TokenRecord> = ctx
-                    .graph
-                    .tokens
-                    .values()
-                    .filter(|t| {
-                        t.raw
-                            .get("name")
-                            .and_then(|v| v.as_object())
-                            .and_then(|n| n.get("property"))
-                            .and_then(|v| v.as_str())
-                            == Some(prop.as_str())
-                    })
-                    .cloned()
-                    .collect();
-                if candidates.is_empty() {
-                    self.active_view = ActiveView::Resolve(ResolveView::new(prop, vec![]));
-                    self.status_message = Some(StatusMessage::info("no match"));
-                    return;
-                }
-                let filtered_graph = TokenGraph::from_records(candidates)
-                    .with_mode_sets(ctx.graph.mode_sets.clone());
-                let mut with_spec: Vec<(&TokenRecord, u32)> = filtered_graph
-                    .tokens
-                    .values()
-                    .map(|t| {
-                        let s = t
-                            .raw
-                            .get("name")
-                            .and_then(|v| v.as_object())
-                            .map(|n| specificity(n, &filtered_graph.mode_sets))
-                            .unwrap_or(0);
-                        (t, s)
-                    })
-                    .collect();
-                with_spec.sort_by(|(a, sa), (b, sb)| {
-                    b.layer
-                        .cmp(&a.layer)
-                        .then_with(|| sb.cmp(sa))
-                        .then_with(|| a.file.cmp(&b.file))
-                        .then_with(|| a.index.cmp(&b.index))
-                });
-                let winner = cascade::resolve(&filtered_graph, &res_ctx);
-                let rows: Vec<ResolvedRow> = with_spec
-                    .iter()
-                    .map(|(t, spec)| {
-                        let value = t
-                            .raw
-                            .get("value")
-                            .map(|v| {
-                                if v.is_string() {
-                                    v.as_str().unwrap_or("").to_string()
-                                } else {
-                                    v.to_string()
-                                }
-                            })
-                            .or_else(|| t.alias_target.clone())
-                            .unwrap_or_default();
-                        let file = t
-                            .file
-                            .file_name()
-                            .map(|f| f.to_string_lossy().into_owned())
-                            .unwrap_or_default();
-                        let is_winner = winner.map(|w| w.name == t.name).unwrap_or(false);
-                        ResolvedRow {
-                            name: display_name(t),
-                            value,
-                            file,
-                            layer: layer_str(t.layer).to_string(),
-                            specificity: *spec,
-                            is_winner,
-                        }
-                    })
-                    .collect();
-                let count = rows.len();
-                self.active_view = ActiveView::Resolve(ResolveView::new(prop, rows));
-                self.status_message =
-                    Some(StatusMessage::info(format!("{count} candidate(s)")));
-            }
-            "describe" | "component" => {
-                if rest.is_empty() {
-                    self.status_message =
-                        Some(StatusMessage::error("describe: component ID required"));
-                    return;
-                }
-                let id = rest.trim();
-                if id.is_empty()
-                    || !id.chars().next().is_some_and(|c| c.is_ascii_lowercase())
-                    || !id
-                        .chars()
-                        .all(|c| c.is_ascii_lowercase() || c.is_ascii_digit() || c == '-')
-                {
-                    self.status_message =
-                        Some(StatusMessage::error(format!("invalid component ID '{id}'")));
-                    return;
-                }
-                let Some(comp_dir) = ctx.components_dir else {
-                    self.status_message = Some(StatusMessage::error(
-                        "describe: no components directory available",
-                    ));
-                    return;
-                };
-                let file_path = comp_dir.join(format!("{id}.json"));
-                if file_path.is_file() {
-                    match std::fs::read_to_string(&file_path) {
-                        Ok(raw_text) => match serde_json::from_str::<serde_json::Value>(&raw_text)
-                        {
-                            Ok(doc) => match serde_json::to_string_pretty(&doc) {
-                                Ok(pretty) => {
-                                    self.active_view = ActiveView::Describe(DescribeView {
-                                        component: id.to_string(),
-                                        pretty_json: pretty,
-                                        scroll: 0,
-                                    });
-                                    self.status_message = None;
-                                }
-                                Err(e) => {
-                                    self.status_message = Some(StatusMessage::error(format!(
-                                        "describe: render error: {e}"
-                                    )));
-                                }
-                            },
-                            Err(e) => {
-                                self.status_message = Some(StatusMessage::error(format!(
-                                    "describe: parse error: {e}"
-                                )));
-                            }
-                        },
-                        Err(e) => {
-                            self.status_message = Some(StatusMessage::error(format!(
-                                "describe: read error: {e}"
-                            )));
-                        }
-                    }
-                } else {
-                    let available: Vec<&str> =
-                        ctx.graph.components.iter().map(|c| c.name.as_str()).collect();
-                    let suggestion = if available.is_empty() {
-                        String::new()
-                    } else {
-                        let prefix_len = id.len().min(3);
-                        let prefix = &id[..prefix_len];
-                        let mut matches: Vec<&str> = available
-                            .iter()
-                            .filter(|&&n| n.starts_with(id))
-                            .copied()
-                            .collect();
-                        if matches.is_empty() {
-                            matches = available
-                                .iter()
-                                .filter(|&&n| n.starts_with(prefix))
-                                .copied()
-                                .collect();
-                        }
-                        if !matches.is_empty() {
-                            format!(
-                                " — did you mean: {}",
-                                matches[..matches.len().min(3)].join(", ")
-                            )
-                        } else {
-                            String::new()
-                        }
-                    };
-                    self.status_message = Some(StatusMessage::error(format!(
-                        "unknown component: {id}{suggestion}"
-                    )));
-                }
-            }
-            "validate" => {
-                let (Some(dataset_path), Some(registry)) =
-                    (ctx.dataset_path, ctx.schema_registry)
-                else {
-                    self.status_message = Some(StatusMessage::error(
-                        "validate: dataset or schema registry unavailable",
-                    ));
-                    return;
-                };
-                match validate::validate_all_with_options_and_names(
-                    dataset_path,
-                    registry,
-                    &HashSet::new(),
-                    ctx.mode_sets_dir,
-                    ctx.components_dir,
-                    None,
-                ) {
-                    Ok(report) => {
-                        let rows: Vec<DiagnosticRow> = report
-                            .errors
-                            .iter()
-                            .map(|d| DiagnosticRow {
-                                severity: "error".to_string(),
-                                rule_id: d.rule_id.clone().unwrap_or_default(),
-                                token: d.token.clone().unwrap_or_default(),
-                                message: d.message.clone(),
-                            })
-                            .chain(report.warnings.iter().map(|d| DiagnosticRow {
-                                severity: "warning".to_string(),
-                                rule_id: d.rule_id.clone().unwrap_or_default(),
-                                token: d.token.clone().unwrap_or_default(),
-                                message: d.message.clone(),
-                            }))
-                            .collect();
-                        let count = rows.len();
-                        self.active_view = ActiveView::Validate(ValidateView::new(rows));
-                        self.status_message =
-                            Some(StatusMessage::info(format!("{count} finding(s)")));
-                    }
-                    Err(e) => {
-                        self.status_message =
-                            Some(StatusMessage::error(format!("validate: {e}")));
-                    }
-                }
-            }
-            "find" => {
-                let fs = FindWizardState::new_with_intent(rest.trim());
-                self.modal = Some(Modal::Find(Box::new(fs)));
-                self.status_message = None;
-            }
-            "name" => {
-                let mut ns = NamingWizardState::new_with_intent(rest.trim());
-                ns.refresh_suggestions(ctx.graph);
-                self.modal = Some(Modal::Naming(Box::new(ns)));
-                self.status_message = None;
-            }
-            "new" | "create" => {
-                let mut ws = WizardState::new_with_intent(rest.trim());
-                ws.refresh_suggestions(ctx.graph);
-                self.modal = Some(Modal::Wizard(Box::new(ws)));
-                self.status_message = None;
-            }
-            other => {
-                self.status_message =
-                    Some(StatusMessage::error(format!("unknown command: {other}")));
-            }
-        }
-    }
+    // ── Palette dispatch — see app_palette.rs ────────────────────────────────
+    // `submit_palette` is defined in a separate `impl App` block in app_palette.rs.
 
     // ── Misc helpers ─────────────────────────────────────────────────────────
 
@@ -1199,14 +608,8 @@ pub(crate) fn save_palette_history(history: &[String]) {
 }
 
 // ── Internal helpers ──────────────────────────────────────────────────────────
-
-pub(crate) fn layer_str(layer: Layer) -> &'static str {
-    match layer {
-        Layer::Foundation => "foundation",
-        Layer::Platform => "platform",
-        Layer::Product => "product",
-    }
-}
+// `layer_str` lives in app_views.rs (re-exported here via `pub use crate::app_views::*`).
+// `apply_scroll_delta` lives in app_views.rs (used by Modal::on_scroll).
 
 /// Advance a `TableState` selection by `delta` rows, clamping at the bounds.
 pub fn move_table_selection(state: &mut TableState, len: usize, delta: i64) {
@@ -1224,16 +627,6 @@ pub(crate) fn rect_contains(rect: Rect, row: u16, col: u16) -> bool {
         && col < rect.x + rect.width
         && row >= rect.y
         && row < rect.y + rect.height
-}
-
-/// Parse the rest-string for `:resolve` into a property name + `ResolutionContext`.
-/// Apply a signed scroll delta to a `u16` scroll position using saturating arithmetic.
-fn apply_scroll_delta(scroll: &mut u16, delta: i32) {
-    if delta > 0 {
-        *scroll = scroll.saturating_add(delta as u16);
-    } else {
-        *scroll = scroll.saturating_sub((-delta) as u16);
-    }
 }
 
 pub(crate) fn parse_resolve_args(rest: &str) -> Result<(String, ResolutionContext), String> {
